pyproject.toml
[project]
name = "camflix"
version = "0.1.0"
description = "Movie recommendation platform inspired by Criticker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "django>=5.2.6",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=21.2.0",

    # Authentication & User Management
    "django-allauth>=0.63.0",

    # Caching & Task Queue
    "redis>=6.4.0",
    "django-redis>=6.0.0",
    "celery>=5.5.3",

    # API & Frontend
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.1",
    "django-htmx>=1.17.3",

    # Environment & Configuration
    "django-environ>=0.11.2",
    "python-dotenv>=1.0.0",

    # Static Files & Media
    "whitenoise>=6.6.0",
    "Pillow>=10.3.0",
    "django-compressor>=4.4",

    # External APIs
    "requests>=2.31.0",
    "httpx>=0.27.0",

    # Data Processing & ML
    "numpy>=1.26.4",
    "pandas>=2.2.2",
    "scikit-learn>=1.4.2",

    # Development Tools
    "django-extensions>=3.2.3",
    "django-debug-toolbar>=4.3.0",
    "ipython>=8.23.0",

    # Security & Monitoring
    "django-ratelimit>=4.1.0",
    "sentry-sdk[django]>=1.45.0",

    # Database Tools
    "django-import-export>=3.3.8",
]
